refactor: Dto에 Serializable 어노테이션 추가
