+++ app/src/main/java/woowacourse/shopping/data/dto/CartDto.kt
@@ -1,15 +1,20 @@
 package woowacourse.shopping.data.dto
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class CartGetResponse(
     val id: Int,
     val quantity: Int,
     val product: ProductGetResponse,
 )
 
+@Serializable
 data class CartAddRequest(
     val productId: Int,
 )
 
+@Serializable
 data class CartPatchRequest(
     val quantity: Int,
 )
