+++ app/src/main/java/woowacourse/shopping/data/dto/OrderDto.kt
@@ -1,20 +1,26 @@
 package woowacourse.shopping.data.dto
 
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class OrderResponse(
+    val orderId: Int,
+    val orderedProducts: List<OrderProductResponse> = listOf(),
+    val payment: PaymentResponse,
+)
+
+@Serializable
 data class OrderPostRequest(
     val orderItems: List<OrderItem>,
     val payment: PaymentRequest,
 )
 
+@Serializable
 data class OrderItem(
     val cartItemId: Int,
 )
 
-data class OrderResponse(
-    val orderId: Int,
-    val orderedProducts: List<OrderProductResponse>,
-    val payment: PaymentResponse,
-)
-
+@Serializable
 data class OrderProductResponse(
     val name: String,
     val price: Int,
