--- app/src/main/java/woowacourse/shopping/data/dto/OrderProductDto.kt
@@ -1,8 +1,0 @@
-package woowacourse.shopping.data.dto
-
-import woowacourse.shopping.domain.model.OrderProduct
-
-fun OrderProduct.toOrderPostInfo(): OrderItem = OrderItem(
-    cartItemId = cartProductId,
-)
-
