+++ app/src/main/java/woowacourse/shopping/data/mapper/OrderMapper.kt
@@ -1,9 +1,9 @@
 package woowacourse.shopping.data.mapper
 
+import woowacourse.shopping.data.dto.OrderItem
 import woowacourse.shopping.data.dto.OrderPostRequest
 import woowacourse.shopping.data.dto.OrderProductResponse
 import woowacourse.shopping.data.dto.OrderResponse
-import woowacourse.shopping.data.dto.toOrderPostInfo
 import woowacourse.shopping.domain.model.Order
 import woowacourse.shopping.domain.model.OrderProduct
 import woowacourse.shopping.domain.model.Price
@@ -12,6 +12,10 @@ import woowacourse.shopping.domain.model.ProductCount
 fun Order.toOrderRequest(): OrderPostRequest = OrderPostRequest(
     orderItems = orderProducts.map { it.toOrderPostInfo() },
     payment = payment.toPaymentRequest(),
+)
+
+fun OrderProduct.toOrderPostInfo(): OrderItem = OrderItem(
+    cartItemId = cartProductId,
 )
 
 fun List<OrderResponse>.toOrders(): List<Order> = map { it.toOrder() }
